anomdtct/requirements.txt
google-cloud-bigquery==1.20.0
pandas==1.0.3
google-cloud-bigquery-storage==0.8.0
pyarrow==0.17.0 
gspread==3.6.0
google-api-python-client==1.8.2
google-auth==1.14.1
